Crate flow

Flow execution and flownode services.

Re-exports

pub use error::Error;

pub use error::Result;

Modules

adapter 🔒

Flow source schema management and stateless streaming execution.

batching_mode 🔒

Run flow as batching mode which is time-window-aware normal query triggered when new data arrives

df_optimizer 🔒

Datafusion optimizer for flow plan

engine 🔒

Define a trait for flow engine, which is used by both streaming engine and batch engine

error

Error definition for flow module

expr 🔒

Small row-batch conversion helper used by stateless streaming.

heartbeat

Send heartbeat from flownode to metasrv

metrics 🔒

Some of the metrics used in the flow module.

repr 🔒

basically a wrapper around the datatype crate for basic Data Representation

server 🔒

Implementation of grpc service for flow node

utils 🔒

Structs

BatchingExecutionGuard

Owns a whole serialized execution round. It may be moved to an execution collaborator, so cancellation of the caller cannot release the round early.

BatchingModeOptions

BatchingTask

ExecuteOnceOutcome

FlowConfig

FlownodeBuilder

FlownodeInstance Builder

FlownodeInstance

The flownode server instance.

FlownodeOptions

FlownodeServer

FlownodeServiceBuilder

Useful in distributed mode

IncrementalAggregateAnalysis

Analysis result for an incremental aggregate plan.

IncrementalAggregateMergeColumn

Describes how one aggregate output field should be merged with the corresponding existing field in the sink table.

PlanInfo

StreamingEngine

TaskArgs

Arguments for creating batching task

TimeWindowExpr

Time window expr like date_bin(INTERVAL '1' MINUTE, ts), this type help with evaluating the expr using given timestamp

Enums

DirtyRestore

FrontendClient

A simple frontend client able to execute sql using grpc protocol

IncrementalAggregateMergeOp

PeerDesc

Describe a peer of frontend

QueryCoverage

Traits

BatchingExecution

BatchingExecutionFactory

GrpcQueryHandlerWithBoxedError

Adapter trait for GrpcQueryHandler that boxes the underlying error into [BoxedError].

Functions

analyze_incremental_aggregate_plan

find_time_window_expr

Return (the column name of time index column, the time window expr, the expected time unit of time index column, the expr's schema for evaluating the time window)

sql_to_df_plan

Convert sql to datafusion logical plan Also support TQL (but only Eval not Explain or Analyze)

Type Aliases

FlowDualEngineRef

Ref to FlowDualEngine

FlowStreamingEngineRef