Module flow::adapter::node_context

Node context, prone to change with every incoming requests

Structs

FlownodeContext

A context that holds the information of the dataflow

IdToNameMap

A tri-directional map that maps table name, table id, and global id

SourceSender

a simple broadcast sender with backpressure and unbound capacity