flow::batching_mode

Module frontend_client

Frontend client to run flow as batching task which is time-window-aware normal query triggered every tick set by user

Structs

DatabaseWithPeer

Enums

FrontendClient

A simple frontend client able to execute sql using grpc protocol

PeerDesc 🔒

Describe a peer of frontend

Traits

GrpcQueryHandlerWithBoxedError

Just like GrpcQueryHandler but use BoxedError

Type Aliases

HandlerMutable 🔒