flow::batching_mode

Module time_window

Time window expr and helper functions

Structs

TimeWindowExpr

Time window expr like date_bin(INTERVAL '1' MINUTE, ts), this type help with evaluating the expr using given timestamp

Constants

DEFAULT_TEST_TIMESTAMP 🔒

Represents a test timestamp in seconds since the Unix epoch.

Functions

binary_search_expr 🔒

calc_expr_time_window_lower_bound 🔒

Find the lower bound of time window in given expr and current timestamp.

columnar_to_ts_vector 🔒

Convert ColumnarValue to Vec<Option<Timestamp>>

create_df_schema_for_ts_column 🔒

eval_phy_time_window_expr 🔒

Expect the phy expression only have one input column with Timestamp type, and also return Timestamp type

find_time_window_expr

Return (the column name of time index column, the time window expr, the expected time unit of time index column, the expr's schema for evaluating the time window)

probe_expr_time_window_upper_bound 🔒

Probe for the upper bound for time window expression

to_phy_expr 🔒