Module flow::compute::render::reduce

Structs

AccumOutput 🔒

Accumulate the output of aggregation functions

ReduceArrange

All arrange(aka state) used in reduce operator

Functions

batch_split_by_key_val 🔒

check_no_future_updates 🔒

eval_distinct_aggrs 🔒

Eval distinct aggregate functions with distinct input arrange

eval_distinct_core 🔒

return distinct rows(distinct by row’s key) from the input, but do not update the arrangement

eval_simple_aggrs 🔒

Eval simple aggregate functions with no distinct input

from_accum_values_to_live_accums 🔒

from_accums_to_offsetted_accum 🔒

convert a list of accumulators to a vector of values with first value as offset of end of each accumulator

from_val_to_slice_idx 🔒

Convert a value to a list of slice index

get_col_diffs 🔒

reduce_accum_subgraph 🔒

eval accumulable reduce plan by eval aggregate function and reduce the result

reduce_batch_subgraph 🔒

reduce_distinct_subgraph 🔒

eval distinct reduce plan, output the distinct, and update the arrangement

reduce_subgraph 🔒

reduce subgraph, reduce the input data into a single row output is concat from key and val

split_rows_to_key_val 🔒

split a row into key and val by evaluate the key and val plan

update_reduce_distinct_arrange 🔒

eval distinct reduce plan, output the distinct, and update the arrangement