flow::df_optimizer

Function sql_to_flow_plan

pub async fn sql_to_flow_plan(
    ctx: &mut FlownodeContext,
    engine: &Arc<dyn QueryEngine>,
    sql: &str,
) -> Result<TypedPlan, Error>

To reuse existing code for parse sql, the sql is first parsed into a datafusion logical plan, then to a substrait plan, and finally to a flow plan.