flow

Module engine

Define a trait for flow engine, which is used by both streaming engine and batch engine

Structs

CreateFlowArgs

The arguments to create a flow

FlowAuthHeader

Traits

FlowEngine

Type Aliases

FlowId

FlowId is a unique identifier for a flow task

TableName