flow

Module error

Error definition for flow module

Structs

ArrowSnafu

SNAFU context selector for the Error::Arrow variant

CacheRequiredSnafu

SNAFU context selector for the Error::CacheRequired variant

ConvertColumnSchemaSnafu

SNAFU context selector for the Error::ConvertColumnSchema variant

CreateFlowSnafu

SNAFU context selector for the Error::CreateFlow variant

DatafusionSnafu

SNAFU context selector for the Error::Datafusion variant

DatatypesSnafu

SNAFU context selector for the Error::Datatypes variant

EvalSnafu

SNAFU context selector for the Error::Eval variant

ExternalSnafu

SNAFU context selector for the Error::External variant

FlowAlreadyExistSnafu

SNAFU context selector for the Error::FlowAlreadyExist variant

FlowNotFoundSnafu

SNAFU context selector for the Error::FlowNotFound variant

FlowNotRecoveredSnafu

SNAFU context selector for the Error::FlowNotRecovered variant

IllegalAuthConfigSnafu

SNAFU context selector for the Error::IllegalAuthConfig variant

IllegalCheckTaskStateSnafu

SNAFU context selector for the Error::IllegalCheckTaskState variant

InsertIntoFlowSnafu

SNAFU context selector for the Error::InsertIntoFlow variant

InternalSnafu

SNAFU context selector for the Error::Internal variant

InvalidQuerySnafu

SNAFU context selector for the Error::InvalidQuery variant

InvalidRequestSnafu

SNAFU context selector for the Error::InvalidRequest variant

JoinTaskSnafu

SNAFU context selector for the Error::JoinTask variant

ListFlowsSnafu

SNAFU context selector for the Error::ListFlows variant

MetaClientInitSnafu

SNAFU context selector for the Error::MetaClientInit variant

NoAvailableFrontendSnafu

SNAFU context selector for the Error::NoAvailableFrontend variant

NotImplementedSnafu

SNAFU context selector for the Error::NotImplemented variant

ParseAddrSnafu

SNAFU context selector for the Error::ParseAddr variant

PlanSnafu

SNAFU context selector for the Error::Plan variant

ShutdownServerSnafu

SNAFU context selector for the Error::ShutdownServer variant

StartServerSnafu

SNAFU context selector for the Error::StartServer variant

SubstraitEncodeLogicalPlanSnafu

SNAFU context selector for the Error::SubstraitEncodeLogicalPlan variant

SyncCheckTaskSnafu

SNAFU context selector for the Error::SyncCheckTask variant

TableNotFoundMetaSnafu

SNAFU context selector for the Error::TableNotFoundMeta variant

TableNotFoundSnafu

SNAFU context selector for the Error::TableNotFound variant

TimeSnafu

SNAFU context selector for the Error::Time variant

UnexpectedSnafu

SNAFU context selector for the Error::Unexpected variant

UnsupportedSnafu

SNAFU context selector for the Error::Unsupported variant

UnsupportedTemporalFilterSnafu

SNAFU context selector for the Error::UnsupportedTemporalFilter variant

Enums

Error

This error is used to represent all possible errors that can occur in the flow module.

Functions

to_status_with_last_err

the outer message is the full error stack, and inner message in header is the last error message that can be show directly to user

Type Aliases

Result

Result type for flow module