flow::expr

Module id

Id is used to identify a dataflow component in plan like Plan::Get{id: Id}, this could be a source of data for an arrangement.

Structs

LocalId

Local id is used in local scope created by Plan::Let{id: LocalId, value, body}

Enums

GlobalId

Global id’s scope is in Current Flow node, and is cross-dataflow

Id

Id is used to identify a dataflow component in plan like Plan::Get{id: Id}