Module flow::expr::relation::accum

Accumulators for aggregate functions that’s is accumulatable. i.e. sum/count

Accumulator will only be restore from row and being updated every time dataflow need process a new batch of rows. So the overhead is acceptable.

Currently support sum, count, any, all and min/max(with one caveat that min/max can’t support delete with aggregate).

Structs

Bool

Bool accumulator, used for Any All Max/MinBool

Float

Accumulates float values for sum over floating numbers.

OrdValue

Accumulates a single Orded Value, useful for min/max aggregations.

SimpleNumber

Accumulates simple numeric values for sum over integer.

Enums

Accum

Accumulates values for the various types of accumulable aggregations.

Traits

Accumulator

Accumulates values for the various types of accumulable aggregations.

Functions

err_try_from_val 🔒

ty_eq_without_precision 🔒

compare type while ignore their precision, including TimeStamp, Time, Duration, Interval