flow::expr

Module utils

This module contains utility functions for expressions.

Functions

find_plan_time_window_expr_lower_bound

Find lower bound for time current in given plan for the time window expr.

find_time_window_lower_bound

Find the lower bound of time window in given expr and current timestamp.