Struct flow::frequency_key_value::FrequencyKeyValue

pub struct FrequencyKeyValue<Key: PartialEq + Eq + Clone, Value: PartialEq + Eq + Clone> {
    // some fields omitted
}

Methods

impl<Key: PartialEq + Eq + Clone, Value: PartialEq + Eq + Clone> FrequencyKeyValue<Key, Value>

fn new() -> FrequencyKeyValue<Key, Value>

fn update(&mut self, key: &Key, value: &Value)

fn sort_by_highest(&mut self) -> Vec<(Key, Vec<(Value, usize)>, usize)>

This returns every unique Key, with a Vec of all the corresponding values encountered, their respecitive count each, and the maximum count encountered for that key over all values.