flow::plan
Module reduce

Structs

AccumulablePlan

Accumulable plan for the execution of a reduction.

AggrWithIndex

Invariant: the output index is the index of the aggregation in full_aggrs which means output index is always smaller than the length of full_aggrs

KeyValPlan

Describe how to extract key-value pair from a Row

Enums

ReducePlan

TODO(discord9): def&impl of Hierarchical aggregates(for min/max with support to deletion) and basic aggregates(for other aggregate functions) and mixed aggregate