flow::recording_rules
Module utils

some utils for helping with recording rule

Structs

AddAutoColumnRewriter

Add to the final select columns like update_at(which doesn’t necessary need to have exact name just need to be a extra timestamp column) and __ts_placeholder(this column need to have exact this name and be a timestamp) with values like now() and 0

AddFilterRewriter

Find out the Filter Node corresponding to innermost(deepest) WHERE and add a new filter expr to it

FindGroupByFinalName

Helper to find the innermost group by expr in schema, return None if no group by expr

Functions

df_plan_to_sql

sql_to_df_plan

Convert sql to datafusion logical plan