flow

Module repr

basically a wrapper around the datatype crate for basic Data Representation

Modules

relation 🔒

Structs

Row

A row is a vector of values.

Constants

BATCH_SIZE

Flow worker will try to at least accumulate this many rows before processing them(if one second haven’t passed)

BROADCAST_CAP

broadcast channel capacity, can be important to memory consumption, since this influence how many updates can be buffered in memory in the entire dataflow TODO(discord9): add config for this, so cpu&mem usage can be balanced and configured by this

SEND_BUF_CAP

The maximum capacity of the send buffer, to prevent the buffer from growing too large

Functions

value_to_internal_ts

Convert a value that is or can be converted to Datetime to internal timestamp

Type Aliases

Diff

System-wide Record count difference type. Useful for capture data change

DiffRow

Default type for a repr of changes to a collection.

Duration

System-wide default duration type, in milliseconds

KeyValDiffRow

Row with key-value pair, timestamp and diff

Timestamp

System-wide default timestamp type, in milliseconds