flow

Module server

Implementation of grpc service for flow node

Structs

FlowService

wrapping flow node manager to avoid orphan rule with Arc<…>

FlownodeBuilder

FlownodeInstance Builder

FlownodeInstance

The flownode server instance.

FlownodeServer

FlownodeServerInner 🔒

FlownodeServerInner is the inner state of FlownodeServer, this struct mostly useful for construct/start and stop the flow node server

FlownodeServiceBuilder

Useful in distributed mode

FrontendInvoker

Basically a tiny frontend that communicates with datanode, different from FrontendClient which connect to a real frontend instead, this is used for flow’s streaming engine. And is for simple query.

Constants

FLOW_NODE_SERVER_NAME

Functions

get_all_flow_ids 🔒

get all flow ids in this flownode

get_flow_auth_options