flow

Module transform

Transform Substrait into execution plan

Modules

aggr 🔒

expr 🔒

literal 🔒

plan 🔒

Macros

not_impl_err 🔒

a simple macro to generate a not implemented error

plan_err 🔒

generate a plan error

Structs

FunctionExtensions

In Substrait, a function can be define by an u32 anchor, and the anchor can be mapped to a name

TumbleFunction

Functions

register_function_to_query_engine

register flow-specific functions to the query engine