flow::transform::aggr

Function find_time_index_in_group_exprs

fn find_time_index_in_group_exprs(group_exprs: &[TypedExpr]) -> Option<usize>

find out the column that should be time index in group exprs(which is all columns that should be keys) TODO(discord9): better ways to assign time index for now, it will found the first column that is timestamp or has a tumble window floor function