flow

Module utils

utilities for managing state of dataflow execution

Structs

ArrangeHandler

A handler to the inner Arrangement, can be cloned and shared, useful for query it’s inner state

Arrangement

A shared state of key-value pair for various state in dataflow execution.

KeyExpiryManager

Determine when should a key expire according to it’s event timestamp in key.

SizeReportSender

Functions

compact_diff_row 🔒

get_value_heap_size

Get a estimate of heap size of a value

Type Aliases

ArrangeReader

Simply a type alias for ReadGuard of Arrangement

ArrangeWriter

Simply a type alias for WriteGuard of Arrangement

Batch

A batch of updates, arranged by key

Spine

A spine of batches, arranged by timestamp TODO(discord9): consider internally index by key, value, and timestamp for faster lookup

StateReportHandler

Handle the size report request, and send the report back