flow::adapter

Module node_context

Node context, prone to change with every incoming requests

Structs

FlownodeContext

A context that holds the information of the dataflow

IdToNameMap

A tri-directional map that maps table name, table id, and global id

SourceSender

a simple broadcast sender with backpressure, bounded capacity and blocking on send when send buf is full note that it wouldn’t evict old data, so it’s possible to block forever if the receiver is slow