flow::compute::render

Module map

Functions

eval_mfp_core 🔒

The core of evaluating MFP operator, given a MFP and a input, evaluate the MFP operator, return the output updates And possibly any number of errors that occurred during the evaluation

mfp_subgraph 🔒