flow::compute

Module types

Structs

Arranged

Arranged is a wrapper around ArrangeHandler that maintain a list of readers and a writer

Collection

A collection, represent a collections of data that is received from a handoff.

CollectionBundle

A bundle of the various ways a collection can be represented.

ErrCollector

A thread local error collector, used to collect errors during the evaluation of the plan

Traits

GenericBundle

Type Aliases

Toff