flow
Module expr

for declare Expression in dataflow, including map, reduce, id and join(TODO!) etc.

Modules

df_func 🔒

Porting Datafusion scalar function to our scalar function to be used in dataflow

error 🔒

Error handling for expression evaluation.

func 🔒

This module contains the definition of functions that can be used in expressions.

id 🔒

Id is used to identify a dataflow component in plan like Plan::Get{id: Id}, this could be a source of data for an arrangement.

linear 🔒

define MapFilterProject which is a compound operator that can be applied row-by-row.

relation 🔒

Describes an aggregation function and it’s input expression.

scalar 🔒

Scalar expressions.

signature 🔒

Function signature, useful for type checking and function resolution.

utils 🔒

This module contains utility functions for expressions.

Structs

Batch

A batch of vectors with the same length but without schema, only useful in dataflow

VectorDiff 🔒

Vector with diff to note the insert and delete

VectorDiffIter 🔒

iterator for VectorDiff

Constants

TUMBLE_END

TUMBLE_START