flow::expr

Module func

This module contains the definition of functions that can be used in expressions.

Macros

generate_binary_signature 🔒

Generate binary function signature based on the function and the input types The user can provide custom signature for some functions in the form of a regular match arm, and the rest will be generated according to the provided list of functions like this:

Structs

BinaryFuncIter

An iterator over the variants of BinaryFunc

Enums

BinaryFunc

BinaryFunc is a function that takes two arguments. Also notice this enum doesn’t contain function arguments, since the arguments are stored in the expression.

UnaryFunc

UnaryFunc is a function that takes one argument. Also notice this enum doesn’t contain function arguments, because the arguments are stored in the expression. (except cast function, which requires a type argument)

UnmaterializableFunc

UnmaterializableFunc is a function that can’t be eval independently, and require special handling

VariadicFunc

VariadicFunc is a function that takes a variable number of arguments.

Statics

SPECIALIZATION 🔒

Functions

add 🔒

and 🔒

div 🔒

get_timestamp_array 🔒

get_ts_as_millisecond 🔒

get_window_start 🔒

mul 🔒

or 🔒

rem 🔒

sub 🔒