Module flow::transform

Transform Substrait into execution plan

Modules

aggr 🔒

expr 🔒

literal 🔒

plan 🔒

Macros

not_impl_err 🔒

a simple macro to generate a not implemented error

plan_err 🔒

generate a plan error

Structs

FunctionExtensions

In Substrait, a function can be define by an u32 anchor, and the anchor can be mapped to a name

TumbleFunction

Constants

TUMBLE_NAME 🔒

Functions

register_function_to_query_engine

register flow-specific functions to the query engine

sql_to_flow_plan

To reuse existing code for parse sql, the sql is first parsed into a datafusion logical plan, then to a substrait plan, and finally to a flow plan.