flow::transform

Module literal

Enums

TimestampPrecision 🔒

Functions

from_bytes 🔒

from_interval_day_sec 🔒

from_interval_year_month 🔒

from_substrait_literal 🔒

Convert a Substrait literal into a Value and its ConcreteDataType (So that we can know type even if the value is null)

from_substrait_type

convert a Substrait type into a ConcreteDataType

to_substrait_literal 🔒

TODO(discord9): this is copy from datafusion-substrait since the original function is not public, will be replace once is exported